fresh remote commit

+++ CRSEG/trackgen.py
@@ -406,5 +406,6 @@ except:
     traceback.print_exc()
     print("some exception has occured!!!!")
     print_no_newline("deleting scratch directory...")
-    shutil.rmtree(scratch_dir)
+    if os.path.exists(scratch_dir):
+        shutil.rmtree(scratch_dir)
     print("exiting")
